+++ core/src/Authorization.html
@@ -1,16 +1,24 @@
 <ul class="fd-menu__list">
-  <li>
-    {#if isLoggedIn}
-      <a href="#" aria-label="Logout" on:click=logout() class="fd-menu__item">Logout</a>
-    {:else}
-      <a href="#" aria-label="Login" on:click=startAuthorization() class="fd-menu__item">Login</a>
-    {/if}
-  </li>
+	<li>
+		{#if isLoggedIn}
+		<a href="#" aria-label="Logout" on:click=logout() class="fd-menu__item">Logout</a>
+		{:else}
+		<a href="#" aria-label="Login" on:click=startAuthorization() class="fd-menu__item">Login</a>
+		{/if}
+	</li>
 </ul>
 
 <script type="text/javascript">
-  import { getConfigValueAsync, getConfigValue, getConfigValueFromObject } from './services/config.js';
-  import { isFunction, isPromise, deepMerge } from './utilities/helpers.js';
+  import {
+    getConfigValueAsync,
+    getConfigValue,
+    getConfigValueFromObject
+  } from './services/config.js';
+  import {
+    isFunction,
+    isPromise,
+    deepMerge
+  } from './utilities/helpers-general.js';
   import { oAuth2ImplicitGrant } from './providers/auth/oAuth2ImplicitGrant.js';
   import { openIdConnect } from './providers/auth/openIdConnect.js';
 
@@ -23,31 +31,40 @@
   };
 
   const getStoredAuthData = () => JSON.parse(localStorage.getItem('luigi.auth'));
-  const newlyAuthorized = () => JSON.parse(localStorage.getItem('luigi.newlyAuthorized'));
-  const isAuthValid = () => getStoredAuthData().accessTokenExpirationDate > Number(new Date());
+  const newlyAuthorized = () =>
+    JSON.parse(localStorage.getItem('luigi.newlyAuthorized'));
+  const isAuthValid = () =>
+    getStoredAuthData().accessTokenExpirationDate > Number(new Date());
 
-  const IdpProviderException = (message) => {
-    return {message, name: 'IdpProviderException'}
+  const IdpProviderException = message => {
+    return { message, name: 'IdpProviderException' };
   };
 
-  const getIdpProviderInstance = async(idpProviderName, idpProviderSettings) => {
+  const getIdpProviderInstance = async (idpProviderName, idpProviderSettings) => {
     if (isFunction(standardProviders[idpProviderName])) {
       return new standardProviders[idpProviderName](idpProviderSettings);
-    };
+    }
 
     // custom provider provided via config:
-    const customIdpProvider = getConfigValueFromObject(idpProviderSettings, 'customIdpProvider');
+    const customIdpProvider = getConfigValueFromObject(
+      idpProviderSettings,
+      'customIdpProvider'
+    );
     if (customIdpProvider) {
-      ['login'].forEach((requiredFnName) => {
+      ['login'].forEach(requiredFnName => {
         if (!isFunction(customIdpProvider[requiredFnName])) {
-          throw new IdpProviderException(`${requiredFnName} function does not exist in custom IDP Provider ${idpProviderName}`);
-        };
+          throw new IdpProviderException(
+            `${requiredFnName} function does not exist in custom IDP Provider ${idpProviderName}`
+          );
+        }
       });
 
       return new customIdpProvider(idpProviderSettings);
-    };
+    }
 
-    throw new IdpProviderException(`IDP Provider ${idpProviderName} does not exist.`);
+    throw new IdpProviderException(
+      `IDP Provider ${idpProviderName} does not exist.`
+    );
   };
 
   export default {
@@ -57,17 +74,17 @@
       const authorizationEnabled = !!idpProviderSettings;
       if (!authorizationEnabled) {
         return;
-      };
+      }
 
       const checkAuth = () => {
         const authData = getStoredAuthData();
         if (!authData || !isAuthValid()) {
           if (getConfigValue('auth.disableAutoLogin')) {
             return;
-          };
+          }
           this.startAuthorization();
           return;
-        };
+        }
 
         this.set({ isLoggedIn: true });
 
@@ -75,30 +92,32 @@
         if (onAuthSuccessfulFn && newlyAuthorized()) {
           localStorage.removeItem('luigi.newlyAuthorized');
           onAuthSuccessfulFn(authData);
-        };
+        }
 
         // TODO: Re-auth via id_token if not logged in
         // TODO: Check periodically if authorization is still valid
         // const onAuthExpiredFn = getConfigValue('auth.events.onAuthExpired');
       };
 
-
-      idpProviderInstance = getIdpProviderInstance(idpProviderName, idpProviderSettings);
+      idpProviderInstance = getIdpProviderInstance(
+        idpProviderName,
+        idpProviderSettings
+      );
       if (isPromise(idpProviderInstance)) {
-        return idpProviderInstance.then((resolved) => {
+        return idpProviderInstance.then(resolved => {
           idpProviderInstance = resolved;
           checkAuth();
         });
-      };
+      }
       checkAuth();
     },
     methods: {
       startAuthorization: function() {
-        idpProviderInstance.login().then((res) => {
+        idpProviderInstance.login().then(res => {
           localStorage.setItem('luigi.newlyAuthorized', true);
           if (res) {
             alert(res);
-          };
+          }
         });
       },
       logout: function() {
@@ -106,26 +125,28 @@
 
         const logoutCallback = () => {
           const onLogoutFn = getConfigValue('auth.events.onLogout');
-          if (onLogoutFn) { onLogoutFn(); };
+          if (onLogoutFn) {
+            onLogoutFn();
+          }
 
           this.set({ isLoggedIn: false });
           localStorage.removeItem('luigi.auth');
         };
 
-        const customLogoutFn = getConfigValue(`auth.${getConfigValue('auth.use')}.logoutFn`);
+        const customLogoutFn = getConfigValue(
+          `auth.${getConfigValue('auth.use')}.logoutFn`
+        );
         if (isFunction(customLogoutFn)) {
           customLogoutFn(idpProviderInstance.settings, authData, logoutCallback);
-        }
-        else if (isFunction(idpProviderInstance.logout)) {
+        } else if (isFunction(idpProviderInstance.logout)) {
           idpProviderInstance.logout(authData, logoutCallback);
-        }
-        else {
+        } else {
           logoutCallback();
           window.location.href = idpProviderInstance.settings.logoutUrl;
-        };
-      },
+        }
+      }
     },
-    components: {},
+    components: {}
     // store: () => store
   };
 </script>
