+++ core/src/utilities/helpers-general.js
@@ -1,4 +1,3 @@
-import { getConfigValue } from '../services/config';
 export const isFunction = anyParam =>
   anyParam && {}.toString.call(anyParam) === '[object Function]';
 
@@ -157,76 +156,3 @@ export const parseParams = paramsString => {
   }
   return result;
 };
-
-export const getNodeParams = params => {
-  const result = {};
-  const paramPrefix = getContentViewParamPrefix();
-  if (params) {
-    Object.entries(params).forEach(entry => {
-      if (entry[0].startsWith(paramPrefix)) {
-        const paramName = entry[0].substr(paramPrefix.length);
-        result[paramName] = entry[1];
-      }
-    });
-  }
-  return result;
-};
-
-export const getPathParams = nodes => {
-  const params = {};
-  nodes
-    .filter(n => n.pathParam)
-    .map(n => n.pathParam)
-    .forEach(pp => {
-      params[pp.key.replace(':', '')] = pp.value;
-    });
-  return params;
-};
-
-const defaultContentViewParamPrefix = '~';
-export const getContentViewParamPrefix = () => {
-  return (
-    getConfigValue('routing.contentViewParamPrefix') ||
-    defaultContentViewParamPrefix
-  );
-};
-
-export const getLastNodeObject = pathData => {
-  const lastElement = [...pathData.navigationPath].pop();
-  return lastElement ? lastElement : {};
-};
-
-export const getDefaultChildNode = function(pathData) {
-  const lastElement =
-    pathData.navigationPath[pathData.navigationPath.length - 1];
-
-  const pathExists = lastElement.children.find(
-    childNode => childNode.pathSegment === lastElement.defaultChildNode
-  );
-
-  if (lastElement.defaultChildNode && pathExists) {
-    return lastElement.defaultChildNode;
-  } else if (lastElement.children && lastElement.children.length > 0) {
-    return lastElement.children[0].pathSegment;
-  } else {
-    return '';
-  }
-};
-
-export const isExistingRoute = function(path, pathData) {
-  if (!path) {
-    return true;
-  }
-
-  const lastElement =
-    pathData.navigationPath[pathData.navigationPath.length - 1];
-  const routeSplit = path.replace(/\/$/, '').split('/');
-  const lastPathSegment = routeSplit[routeSplit.length - 1];
-
-  return lastElement.pathSegment === lastPathSegment;
-};
-
-export const buildRoute = (node, path, params) =>
-  !node.parent
-    ? path + (params ? '?' + params : '')
-    : buildRoute(node.parent, `/${node.parent.pathSegment}${path}`, params);
